Extract register email page from main app file
